+++ src/main/java/pl/coderslab/repository/WeekdayRepository.java
@@ -3,8 +3,6 @@ package pl.coderslab.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.coderslab.model.Weekday;
 
-import java.util.List;
-
-public interface WeekdayRepository extends JpaRepository<Weekday, Long> {
+public interface WeekdayRepository extends JpaRepository<Weekday, Integer> {
 
 }
